# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentCitationResponse": {
        "properties": {
          "citation_id": {
            "title": "Citation Id",
            "type": "string"
          },
          "excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excerpt"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "knowledge_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Knowledge Id"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          }
        },
        "required": [
          "citation_id",
          "label",
          "kind"
        ],
        "title": "AgentCitationResponse",
        "type": "object"
      },
      "AgentConversationListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AgentConversationSummaryResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "AgentConversationListResponse",
        "type": "object"
      },
      "AgentConversationResponse": {
        "properties": {
          "conversation_id": {
            "format": "uuid",
            "title": "Conversation Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "research_map": {
            "additionalProperties": true,
            "title": "Research Map",
            "type": "object"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "turn_count": {
            "title": "Turn Count",
            "type": "integer"
          },
          "turns": {
            "items": {
              "$ref": "#/components/schemas/AgentTurnResponse"
            },
            "title": "Turns",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "conversation_id",
          "title",
          "updated_at",
          "turn_count",
          "created_at",
          "turns"
        ],
        "title": "AgentConversationResponse",
        "type": "object"
      },
      "AgentConversationSummaryResponse": {
        "properties": {
          "conversation_id": {
            "format": "uuid",
            "title": "Conversation Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "turn_count": {
            "title": "Turn Count",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "conversation_id",
          "title",
          "updated_at",
          "turn_count"
        ],
        "title": "AgentConversationSummaryResponse",
        "type": "object"
      },
      "AgentMessageResponse": {
        "properties": {
          "citations": {
            "items": {
              "$ref": "#/components/schemas/AgentCitationResponse"
            },
            "title": "Citations",
            "type": "array"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "message_id": {
            "format": "uuid",
            "title": "Message Id",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          }
        },
        "required": [
          "message_id",
          "role",
          "content",
          "sequence",
          "created_at"
        ],
        "title": "AgentMessageResponse",
        "type": "object"
      },
      "AgentToolTraceResponse": {
        "properties": {
          "call_id": {
            "title": "Call Id",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "input": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input"
          },
          "output": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Output"
          },
          "phase": {
            "enum": [
              "started",
              "finished",
              "failed"
            ],
            "title": "Phase",
            "type": "string"
          },
          "tool": {
            "title": "Tool",
            "type": "string"
          }
        },
        "required": [
          "tool",
          "phase",
          "call_id"
        ],
        "title": "AgentToolTraceResponse",
        "type": "object"
      },
      "AgentTurnRequest": {
        "properties": {
          "conversation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "document_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "document_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Version"
          },
          "message": {
            "maxLength": 12000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "section_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Id"
          },
          "task_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "theory_plan_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Theory Plan Id"
          },
          "workspace": {
            "default": "agent",
            "enum": [
              "agent",
              "research"
            ],
            "title": "Workspace",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "AgentTurnRequest",
        "type": "object"
      },
      "AgentTurnResponse": {
        "properties": {
          "assistant": {
            "$ref": "#/components/schemas/AgentMessageResponse"
          },
          "canvas_patches": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Canvas Patches",
            "type": "array"
          },
          "knowledge_release_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Knowledge Release Id"
          },
          "tool_traces": {
            "items": {
              "$ref": "#/components/schemas/AgentToolTraceResponse"
            },
            "title": "Tool Traces",
            "type": "array"
          },
          "turn_id": {
            "format": "uuid",
            "title": "Turn Id",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AgentMessageResponse"
          }
        },
        "required": [
          "turn_id",
          "user",
          "assistant"
        ],
        "title": "AgentTurnResponse",
        "type": "object"
      },
...
      "ErrorCode": {
        "enum": [
          "unauthenticated",
          "session_expired",
          "not_found",
          "method_not_allowed",
          "research_task_not_found",
          "validation_error",
          "phenomenon_unconfirmed",
          "no_adopted_theory",
          "candidate_ineligible",
          "external_candidate_adoption_blocked",
          "model_timeout",
          "no_reliable_candidate",
          "insufficient_sources",
          "stale_framework_revision",
          "unresolved_blocking_audit",
          "not_implemented",
          "internal_server_error"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "trace_id"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "群学致知前后端架构基线 API。",
    "title": "群学致知 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/agent/conversations": {
      "get": {
        "operationId": "list_agent_conversations",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentConversationListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "List Agent Conversations",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/agent/conversations/{conversation_id}": {
      "get": {
        "operationId": "get_agent_conversation",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentConversationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Agent Conversation",
        "tags": [
          "agent"
        ]
      }
    },
    "/api/agent/turns": {
      "post": {
        "operationId": "stream_agent_turn",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentTurnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Server-sent Agent events"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Stream Agent Turn",
        "tags": [
          "agent"
        ]
      }
    },
...
  }
}